requirements.txt
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.7.2
astroid==2.9.2
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
black==20.8b1
bleach==4.1.0
Brotli==1.1.0
cffi==1.14.6
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.4
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.7
django-admin-interface==0.26.0
django-adminlte-3==0.1.6
django-ckeditor==6.2.0
django-colorfield==0.9.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-js-asset==1.2.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.9.1
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fonttools==4.53.0
greenlet==2.0.2
gunicorn==22.0.0
html5lib==1.1
importlib-metadata==6.0.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.7.1
lxml==5.2.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
openpyxl==3.1.2
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.11.2
pdfminer.six==20231228
pdfquery==0.4.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.20.3
psycopg2==2.9.7
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.20
pydyf==0.10.0
Pygments==2.10.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==2.4.7
pyphen==0.15.0
pyquery==2.0.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2021.3
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
regex==2023.10.3
roman==4.2
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==2.0.9
sqlparse==0.4.4
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tinycss2==1.3.0
tk==0.1.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typed-ast==1.5.5
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.21.0
waitress==3.0.0
wcwidth==0.2.5
weasyprint==62.2
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.13.3
zipp==3.15.0
zopfli==0.2.3
